Add files via upload
Uploaded version in development.
The two main pieces of code currently are in the SECDataMiner.py file. This will scrape the SEC database and find the company you are searching for. It then exports the history of the company to a reports folder, sorted by year. There are still some bugs to be worked out. This utilizes BeautifulSoup for parsing html, Pandas for data frames, and XlsxWriter for writing to Excel documents.
The other piece is in the LoginFrame.py file and this is the first part of the GUI module.